refactor: split LCD into dedicated package
