+++ apps/web/package.json
@@ -5,6 +5,7 @@
   "type": "module",
   "dependencies": {
     "@z80emu/assembler-z80": "file:../../packages/assembler-z80",
+    "@z80emu/lcd-144x32": "file:../../packages/lcd-144x32",
     "@z80emu/machine-pcg815": "file:../../packages/machine-pcg815"
   },
   "devDependencies": {
