+++ apps/web/src/main.ts
@@ -1,17 +1,19 @@
 import {
-  getGlyphForCode,
-  hasGlyphForCode,
   KEY_MAP,
   KEY_MAP_BY_CODE,
+  PCG815Machine,
+  MONITOR_PROMPT_RESUME_ADDR
+} from '@z80emu/machine-pcg815';
+import { assemble } from '@z80emu/assembler-z80';
+import {
+  decodeMachineText,
+  getGlyphForCode,
+  hasGlyphForCode,
   LCD_GLYPH_HEIGHT,
   LCD_GLYPH_WIDTH,
   LCD_HEIGHT,
-  LCD_WIDTH,
-  PCG815Machine,
-  MONITOR_PROMPT_RESUME_ADDR,
-  decodeMachineText
-} from '@z80emu/machine-pcg815';
-import { assemble } from '@z80emu/assembler-z80';
+  LCD_WIDTH
+} from '@z80emu/lcd-144x32';
 
 import './styles.css';
