+++ apps/web/vitest.config.ts
@@ -7,6 +7,7 @@ export default defineConfig({
     alias: {
       '@z80emu/core-z80': path.resolve(__dirname, '../../packages/core-z80/src/index.ts'),
       '@z80emu/firmware-monitor': path.resolve(__dirname, '../../packages/firmware-monitor/src/index.ts'),
+      '@z80emu/lcd-144x32': path.resolve(__dirname, '../../packages/lcd-144x32/src/index.ts'),
       '@z80emu/machine-chipsets': path.resolve(__dirname, '../../packages/machine-chipsets/src/index.ts'),
       '@z80emu/machine-pcg815': path.resolve(__dirname, '../../packages/machine-pcg815/src/index.ts')
     }
