+++ packages/lcd-144x32/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@z80emu/lcd-144x32",
+  "version": "0.1.0",
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "test": "vitest run",
+    "typecheck": "tsc -p tsconfig.json --noEmit"
+  }
+}
