+++ packages/machine-pcg815/src/index.ts
@@ -1,8 +1,6 @@
 // machine-pcg815 パッケージの公開 API 入口。
-export * from './font5x7';
 export * from './hardware-evidence';
 export * from './hardware-map';
 export * from './keyboard-map';
-export * from './lcd-text';
 export * from './machine';
 export * from './types';
