+++ packages/machine-pcg815/tests/char-function.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import { PCG815Machine, decodeMachineText } from '../src';
+import { decodeMachineText } from '@z80emu/lcd-144x32';
+import { PCG815Machine } from '../src';
 
 function encode(lines: readonly string[]): number[] {
   const out: number[] = [];
@@ -34,4 +35,3 @@ describe('z80 basic CHAR$/CHR$ print', () => {
     expect(screen).toContain('AB');
   });
 });
-
