+++ packages/machine-pcg815/tests/randomize-rnd.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import { PCG815Machine, decodeMachineText } from '../src';
+import { decodeMachineText } from '@z80emu/lcd-144x32';
+import { PCG815Machine } from '../src';
 
 function encode(lines: string[]): number[] {
   const out: number[] = [];
